Refaturando arquivo do grupo.yml se a versão do plugin for inferior a 4.0.0 ler o arquivo grupo.yml e salva no banco de dados

+++ src/main/java/mc/elderbr/smarthopper/file/Config.java
@@ -40,8 +40,8 @@ public class Config {
     }
 
     //=================== VERSAO DO PLUGIN ===============================/
-    public static double VERSION() {
-        return Double.parseDouble(YML.getString("version").replaceAll("[.]", ""));
+    public static int VERSION() {
+        return Integer.parseInt(YML.getString("version").replaceAll("[.]", ""));
     }
 
     public static void SET_VERSION(String version) {
